Regularize the bitcount for the private key.
This is so that attackers cannot use a side channel attack to learn the number of leading zeros. This causes about 1% slower keypair generation and ECDH.

Note: this is not done for secp160r1, since it would have a larger performance impact (about 2% slower on average) and requires the vli_xxx_n functions, leading to a significant increase in code size.