Some work on fixing ARM asm for Xcode compiler
The Xcode compiler (some version of clang, I guess) doesn't handle adc
instructions properly. Specifically it can't handle the 2-argument version
if one of the registers is a hi reg and it is in Thumb mode; it thinks that
it needs to use a 16-bit instruction only, and errors out. By explicitly
using the 3-argument version of adc, we can trick it into working correctly.